# Remove unnecessary unsafe blocks